Add feature to export conversations
The button appears on conversations. If the button is in a folder, a folder download button appears.

A simple text file is emitted. For a folder a title for the folder is appended and the conversation follows.

Currently it's a plain text message. Potentially in the future a word document could be emitted to show images. Or maybe a zip file.