Add dates to messages
I chose string for the date type as it works better with json exporting. There's not really a native datetime type in javascript.

Note that Azure OpenAI doesn't expect "timestamp" so I had to make a OpenAIMessage type and strip the timestamp before posts are made there.

Using format and date-fns for timezone aware dates.