Support gpt-5 style citations
It seems to come from the file_search tool definition described here: https://community.openai.com/t/gpt-5-not-quoting-excerpts-with-quotation-marks-in-the-response-api/1365308/4

The format is a bit strange and comes down in several tokens. My solution is to "vacuum" (combine) a fixed 50 chunks back to get the whole citation. Then the citation is converted to the gpt-4 format. This then uses the existing code to convert it to a readable format.

I assume the files are in order, for example file0 corresponds to the first file in file index. I will have to test this.